requirements.txt
Django>=3.2,<3.3
wagtail>=2.14,<2.15
django-extensions==3.2.9
django-widget-tweaks==1.4.9
wagtailcodeblock-1.25.0.2
